isServiceUp must honor --no-check
