+++ pkg/cli/cc/cc.go
@@ -4,11 +4,16 @@ import (
 	"context"
 	"fmt"
 	"os"
+	"time"
 
 	"github.com/aserto-dev/clui"
 	"github.com/aserto-dev/topaz/pkg/cli/cc/iostream"
 	"github.com/aserto-dev/topaz/pkg/cli/dockerx"
+	"github.com/fullstorydev/grpcurl"
 	"github.com/samber/lo"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/credentials"
+	"google.golang.org/grpc/credentials/insecure"
 )
 
 type CommonCtx struct {
@@ -49,4 +54,29 @@ func (c *CommonCtx) CheckRunStatus(containerName string, expectedStatus runStatu
 	}
 
 	return lo.Ternary(running, StatusRunning, StatusNotRunning) == expectedStatus
+}
+
+func (c *CommonCtx) IsServing(grpcAddress string) bool {
+	if c.NoCheck {
+		return false
+	}
+
+	tlsConf, err := grpcurl.ClientTLSConfig(true, "", "", "")
+	if err != nil {
+		return false
+	}
+
+	creds := credentials.NewTLS(tlsConf)
+
+	opts := []grpc.DialOption{
+		grpc.WithUserAgent("topaz/dev-build (no version set)"),
+		grpc.WithTransportCredentials(insecure.NewCredentials()),
+	}
+
+	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
+	defer cancel()
+
+	_, err = grpcurl.BlockingDial(ctx, "tcp", grpcAddress, creds, opts...)
+
+	return err == nil
 }
